+++ src/Stack/Config.hs
@@ -331,7 +331,8 @@ configFromConfigMonoid
          configModifyCodePage = fromFirstTrue configMonoidModifyCodePage
          configRebuildGhcOptions = fromFirstFalse configMonoidRebuildGhcOptions
          configApplyGhcOptions = fromFirst AGOLocals configMonoidApplyGhcOptions
-         configAllowNewer = fromFirst AllowNewerNone configMonoidAllowNewer
+         configAllowNewer = fromFirst False configMonoidAllowNewer
+         configAllowNewerDeps = coerce configMonoidAllowNewerDeps
          configDefaultTemplate = getFirst configMonoidDefaultTemplate
          configDumpLogs = fromFirst DumpWarningLogs configMonoidDumpLogs
          configSaveHackageCreds = fromFirst True configMonoidSaveHackageCreds